Add explicit zero type check
In certain circumstances (such as a definition which has an unannotated
mailbox type, which is not called), a pattern variable p will not have
an upper bound. In this circumstance, 0 will be inferred, which is an
invalid type. This patch adds an explicit check to ensure that an
inferred pattern solution is usable (i.e., contains no 0 types)

It also updates the test suite to the new layout.