Replace 'free' sugar with 'empty' guard and 'free' construct (#13)
Before, we were using the syntactic sugar:

[[ free V ]] = guard V : 1 { free -> () }

For Mailboxer we actually need a slightly more general solution:

Replace the 'free' guard with an 'empty(x) -> M' guard, which is triggered in the same scenarios as 'free' but does not deallocate the variable. Instead the mailbox is bound to 'x' with type ?1.
Delete the free V sugar
Introduce a native free V : ?1 -> Unit construct that explicitly deallocates a mailbox of type ?1
Everything typable before is typable now.