Generalise pairs to n-tuples (#19)
This (surprisingly annoying-to-write?) PR generalises the pairs that we had previously to tuples of arbitrary length. This will be useful for the multiple mailboxes extension.

Note that currently we still can't do nested pattern matching: while we can construct `((1, 2), 3)`, we can't deconstruct it like `let ((x, y), z) = ((1, 2), 3) in y` just yet. One for another patch.

This patch also removes the unit constant and unit type (which are nullary tuples and the nullary tuple type respectively). 
To preserve backwards compatibility, `Unit` is now an alias for `()`.